refactor: Move Member and MemberRepository into domain package
