--- jpashop/src/main/java/jpabook/jpashop/Member.java
@@ -1,19 +1,0 @@
-package jpabook.jpashop;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Getter
-@Setter
-public class Member {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-    private String username;
-}
