+++ jpashop/src/main/java/jpabook/jpashop/domain/Member.java
@@ -1,0 +1,27 @@
+package jpabook.jpashop.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+public class Member {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "member_id")
+    private Long id;
+
+    private String username;
+
+    @Embedded
+    private Address address;
+
+    @OneToMany(mappedBy = "member") // read only
+    private List<Order> orders = new ArrayList<>();
+}
