+++ jpashop/src/main/java/jpabook/jpashop/domain/MemberRepository.java
@@ -1,4 +1,4 @@
-package jpabook.jpashop;
+package jpabook.jpashop.domain;
 
 import org.springframework.stereotype.Repository;
